backend/api_foodgram/requirements.txt
Django==2.2.16
django-filter
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser
Pillow==9.0.0
psycopg2-binary==2.8.6
python-dotenv